refactor: extract SimilarIssueResolution.swift from IssueExportReview.swift
Splits the `SimilarIssueResolution` enum out.

Byte-identical move of lines 3-20. Both files import Foundation.

IssueExportReview.swift: 130 → 111 lines. Tests: 667.
